feat + test: add infinite money glitch + relevant tests
infinite money glitch = mining reward pool empties less quickly as the pool gets smaller. This is done by dishing out rewards as a percentage of the reward pool

+++ contracts/libraries/storage/LibSearchSECORewardingStorage.sol
@@ -18,7 +18,7 @@ library LibSearchSECORewardingStorage {
         uint hashReward;
 
         /// @notice Defines the percentage of the pool that is paid out to the miner.
-        uint32 miningRewardPoolPayoutRatio; // ppm
+        int128 miningRewardPoolPayoutRatio; // 64.64
     }
 
     function getStorage() internal pure returns (Storage storage ds) {
